Format and mount disks

+++ install.sh
@@ -5,7 +5,7 @@ kernel_version_file=$kernel_version.tar.gz
 image_file=ft_linux.img  # It could be replaced with local disk, but for obvious reasons (avoid accidents, no need for a VM), I prefer to work on a image.
 
 echo Installing needed package
-sudo pacman -Sy --needed gcc bc parted || exit 1
+sudo pacman -Sy --needed bc parted base-devel wget || exit 1
 
 
 if [ ! -f $kernel_version_file ]; then
@@ -36,10 +36,44 @@ make || exit 1
 
 cd ..
 
-#if [ ! -f $image_file ]; then
+if [ ! -f $image_file ]; then
+  echo Creating the image disk
+
   dd if=/dev/zero of=$image_file bs=1M count=8192 || exit 1
   parted $image_file mktable msdos || exit 1
   parted $image_file mkpart primary fat32 1 1024 || exit 1
   parted $image_file mkpart primary ext4 1024 7167 || exit 1
   parted $image_file mkpart primary linux-swap 7168 8191 || exit 1
-#fi
+
+fi
+
+loopdisk=$(losetup -j ./ft_linux.img | head -n 1 | cut -f1 -d":")
+if [ -z "$loopdisk" ]; then
+  loopdisk=$(sudo losetup -Pf --show ft_linux.img) || exit 1
+fi
+
+echo "Loopdisk at $loopdisk"
+
+echo "Formatting and mounting..."
+
+if [ -d ./boot ]; then
+  sudo umount ./boot
+  sudo umount ./system
+  sudo umount ./swap
+fi
+
+mkdir -p boot system swap
+
+export LFS=$(pwd)/system
+export BFS=$(pwd)/boot
+umask 022 # https://www.linuxfromscratch.org/lfs/view/stable/chapter02/aboutlfs.html
+
+sudo mkfs.vfat "${loopdisk}p1"
+sudo mkfs.ext4 "${loopdisk}p2"
+sudo mkswap "${loopdisk}p3"
+sudo mount -w "${loopdisk}p1" $BFS
+sudo mount -w "${loopdisk}p2" $LFS
+
+sudo mkdir -p $BFS/efi $LFS/home $LFS/usr $LFS/opt $LFS/usr/src
+
+
